rename of service file as per review comments
